Copy Dockerfile commands to Singularity file
The commands related to the user fenics inside the docker container
works poorly in Singularity and there are some error messages. Lets
try to use the standard ubuntu docker image to bootstrap from
instead. This is probably tested, and we can just copy the commands
over from FEniCS docker repo of dockerfiles